Sorting an Array in Descending Order
In this program, we need to sort the given array in descending order such that elements will be arranged from largest to smallest. This can be achieved through two loops. Outer loop will select a element and inner loop allow us to compare selected element with rest of the elements.